delete old files (new.js and related example)
