+++ example_write.js
@@ -1,9 +1,13 @@
-var prov=require('./dist/new');
+var Anvil=require("./dist/anvil");
+var Vec3 = require("vec3");
 
+var chunk = new Anvil(process.argv[2] ? process.argv[2] : "world/lttp");
 
-prov.setChunk(process.argv[2] ? process.argv[2] : "world/lttp",{"name":"","type":"compound","value":{}}, 0, 0).then(function (data) {
-    console.log(data);
+var c=chunk.save(32,0,{"name":"","type":"compound","value":{}});
+
+c.then(function(){
+    console.log("saved");
   })
-  .catch(function (err) {
+  .catch(function(err){
     console.log(err.stack);
   });
